New docker compose file and updated README
Docker compose file was added to the project to provide an easier way of setting up and managing the database container. It provides a concise and simple syntax which makes it the perfect tool for the job. Also, the README file was updated to describe how to perform the initial set up of the database container using docker compose.